requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
openai==1.12.0
pandas==2.2.1 